Add functionality to stream functions and code refactoring
