+++ src/streams/write.js
@@ -1,5 +1,14 @@
+import fs from 'fs'
+import path from 'path'
+import { getFilePath } from '../tools/filepath.js'
 const write = async () => {
-    // Write your code here 
-};
+	const { __dirname } = getFilePath(import.meta.url)
+	const fileToWritePath = path.join(__dirname, 'files', 'fileToWrite.txt')
+	const writeStream = fs.createWriteStream(fileToWritePath)
+	process.stdin.on('data', (data) => {
+		writeStream.write(data.toString())
+		process.exit()
+	})
+}
 
-await write();
+await write()
